Use byte instead of int backing type for cancelation state.

+++ ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/Cancelations/Internal/CancelationInternal.cs
@@ -124,7 +124,7 @@ namespace Proto.Promises
                 }
             }
 
-            internal enum State : int
+            internal enum State : byte
             {
                 Pending,
                 Disposed,
